차량등록기능구현

+++ src/main/java/com/example/helloCar/domain/hellocar/repository/HelloCarRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HelloCarRepository extends JpaRepository<HelloCar, Long> {
+public interface HelloCarRepository extends JpaRepository<HelloCar, Integer> {
 }
